Flat rides: allow >255px redraw bounds for tall high-res sprites
PROBLEM
A large custom flat ride (the ~648px Ferris Wheel) had its upper ~1/3 clipped as it
animated - the top showed the tiles behind it. Root cause: the redraw (invalidation)
region is bounded by EntityBase::EntitySpriteData.{width,heightMin,heightMax}, which are
uint8_t (255px max). RCT2's own sprites are tiny (~100px) so base rides never approached
this; our high-res art is ~5x larger and blows past it, so the frame-change invalidation
never marks the sprite's top dirty and it isn't redrawn.

WHY NOT JUST WIDEN EntitySpriteData
Those fields are SERIALIZED into the park save (ParkFile.cpp ReadWriteEntityCommon), so
widening them to uint16 would break the save format (version bump + migration). Avoided.

FIX (save-safe, contained to the generic flat ride)
ViewportsInvalidate() already takes int32 bounds - only the entity's uint8 STORAGE caps
us. So:
- FlatRideRotationDescriptor.Invalidation{HalfWidth,HeightAbove,HeightBelow}: uint8 ->
  uint16. The descriptor is rebuilt from the ride object's JSON every load, NOT stored in
  the park save, so this changes no save format.
- RideObject.cpp parses them as uint16.
- Vehicle::UpdateFlatRideGeneric, on a frame change, invalidates the FULL descriptor-sized
  region DIRECTLY via ViewportsInvalidate (int32) instead of the entity's capped
  invalidate(); falls back to invalidate() when the descriptor sets no explicit size.
- GenericFlatRide still writes the (uint8, serialized) EntitySpriteData for the
  selection/placement rect, now clamped to 255 - the animation redraw no longer depends on
  it. (Selecting the extreme top of a >255px ride is the one minor cost; the visual is
  correct.)

Authored via AttractionEditorTool (invalidation bounds now emitted as uint16).

+++ src/openrct2/object/RideObject.cpp
@@ -1267,9 +1267,11 @@ namespace OpenRCT2
         data->Descriptor.BbLengthX               = Json::GetNumber<uint8_t>(j["bbLengthX"], 24);
         data->Descriptor.BbLengthY               = Json::GetNumber<uint8_t>(j["bbLengthY"], 24);
         data->Descriptor.BbLengthZ               = Json::GetNumber<uint8_t>(j["bbLengthZ"], 48);
-        data->Descriptor.InvalidationHalfWidth   = Json::GetNumber<uint8_t>(j["invalidationHalfWidth"]);
-        data->Descriptor.InvalidationHeightAbove = Json::GetNumber<uint8_t>(j["invalidationHeightAbove"]);
-        data->Descriptor.InvalidationHeightBelow = Json::GetNumber<uint8_t>(j["invalidationHeightBelow"]);
+        // uint16 (not uint8): a large high-res sprite needs a redraw region past 255px - see
+        // FlatRideRotationDescriptor and Vehicle::UpdateFlatRideGeneric's direct invalidation.
+        data->Descriptor.InvalidationHalfWidth   = Json::GetNumber<uint16_t>(j["invalidationHalfWidth"]);
+        data->Descriptor.InvalidationHeightAbove = Json::GetNumber<uint16_t>(j["invalidationHeightAbove"]);
+        data->Descriptor.InvalidationHeightBelow = Json::GetNumber<uint16_t>(j["invalidationHeightBelow"]);
         data->Descriptor.RiderFrameStride        = Json::GetNumber<uint8_t>(j["riderFrameStride"]);
         // Quarter-turns to rotate the sprite relative to the footprint (0-3); masked so a
         // stray value can't index outside the 4 direction blocks.
